Fix beans.xml files for CDI Lite
Quarkus complains that bean-discovery-mode="all" is not supported. This
is because CDI Lite (on which Quarkus is based) only supports
"annotated".

With this change, only the modules where CDI beans are actually defined
keep their beans.xml file and the discovery mode is changed to
"annotated". Some beans.xml files have been moved to test resources
since they are needed by Arquillian.